DATAJDBC-95 - Saving and loading a trivial entity.
Creation of the necessary sql statements is now dynamic and driven in a
trivial way by the entity.

Known issues with the solution that need to get fixed later:

Sql generating code is in the repository and should go somewhere else.

Mapping logic is very trivial and should go in a separate place.