DATAJDBC-233 - Adapted internal APIs to support immutable entities.
!!! WARNING: Do not merge! Does not compile yet!!!

Changed JdbcAggregateOperations to return an instance from its ….save(…) method, so that potential changes of properties that return in the instance being exchanged for immutable entities can be reflected in clients.

Moved downstream APIs to PersistentPropertyAccessor as this transparently exchanges the captured instances so that the result of a manipulation can be obtained later from the level on which the PersistentPropertyAccessor has been created. AggregateChange now operates on that PersistentPropertyAccessor API.

TODO: the code to handle related entities is quite convoluted and I didn't manage to wrap my head around how to actually make it work with PersistentPropertyAccessor yet. Ideally we change the code to calculate PersistentPropertyPath instances so that it can rather operate on PersistentPropertyAccessor.setProperty(PersistentPropertyPath, Object) level as that will make sure that nested immutable objects trigger a replacement to bubble up to the root aggregate.