DATAJDBC-326 - Support conversion of backreferences.
Ids used as backreferences now get properly converted.

Introduced ParentKeys.
ParentKeys hold information about data that needs to be considered for inserts or updates but is not part of the entity.
Appart from column names and values they also hold information about the desired JdbcType in order to facilitate conversions.
This replaces the Map handed around in the past.