#291 - Adapt to SqlIdentifier changes in Spring Data Relational.
We now use SqlIdentifier to encapsulate the table and column name across the API. The fluent DatabaseClient API accepts the table name and select column projections as SqlIdentifier. The Criteria API remains String-based as names for Query and Update objects may reference either column names or property names.

QueryMapper and UpdateMapper require now a R2DBC dialect object to be constructed. R2dbcMappingContext is configured in a way to not require quoting to preserve backwards compatibility.