DATAJDBC-455 - Adds dynamic Dialect detection.
So far the user had to specify an `Dialect` themselves if they wanted to use anything but HSQLDB.
We now identify the supported databases and pick the appropriate `Dialect`.

Vendors who want to offer support for their database may provide an implementation of `JdbcDriverProvider` and register it using a  file under the key `org.springframework.data.jdbc.repository.config.JdbcDialectResolver$JdbcDialectProvider`.