On join conditions (from, to) we should use Column and not Alias Column.
if we use Alias column on Join Conditions we will have alias used, what will cause an error.

We should not have `join b on a.col = b.alias`. b.alias should not be present in the join, it will raise an error.

The correct version should be `join b on a.col = b.col`.