`StringBuilder` can be replaced with String concatenation.
In this case the usage of `StringBuilder` has no additional benefit, since it is not inside a loop where the jvm may have problems detecting String concatenation that can be optimized. The usage of "simple" String concatenation makes the code more readable and the code will be automatically optimized by the compiler. (So the final result may use `StringBuilder`).