Add failsafe:verify goal as an execution.
The failsafe plugin only fails during it's verify goal, not when actually executing the tests.

The goal is by default added to the verify phase.
But for the database tests this adds it to the integration-test phase.
This ensures that the build fail as soon a test for one database fails.

Closes #1637